jwt access done; backend api now checks for user permissions before executing queires

+++ server/datasources/instruments.js
@@ -1,7 +1,6 @@
 /* eslint-disable max-len */
 // This file deals with what methods a model model should have
 const { DataSource } = require('apollo-datasource');
-const { responsePathAsArray } = require('graphql');
 const SQL = require('sequelize');
 
 function validateInstrument({
@@ -44,6 +43,11 @@ class InstrumentAPI extends DataSource {
    */
   initialize(config) {
     this.context = config.context;
+  }
+
+  checkPermission() {
+    const { user } = this.context;
+    return user.isAdmin || user.instrumentPermission;
   }
 
   async countAllInstruments() {
@@ -419,6 +423,10 @@ class InstrumentAPI extends DataSource {
     const validation = validateInstrument({
       modelNumber, vendor, serialNumber, comment, assetTag,
     });
+    if (!this.checkPermission()) {
+      response.message = 'ERROR: User does not have permission.';
+      return JSON.stringify(response);
+    }
     if (!validation[0]) {
       // eslint-disable-next-line prefer-destructuring
       response.message = validation[1];
@@ -501,6 +509,10 @@ class InstrumentAPI extends DataSource {
     const response = { message: '', success: false };
     const storeModel = await this.store;
     this.store = storeModel;
+    if (!this.checkPermission()) {
+      response.message = 'ERROR: User does not have permission.';
+      return JSON.stringify(response);
+    }
     await this.store.instruments.destroy({ where: { id } });
     await this.store.calibrationEvents.destroy({
       where: { calibrationHistoryIdReference: id },
@@ -517,6 +529,10 @@ class InstrumentAPI extends DataSource {
     const validation = validateInstrument({
       modelNumber, vendor, serialNumber, comment, assetTag,
     });
+    if (!this.checkPermission()) {
+      response.message = 'ERROR: User does not have permission.';
+      return JSON.stringify(response);
+    }
     if (!validation[0]) {
       // eslint-disable-next-line prefer-destructuring
       response.message = validation[1];
@@ -612,6 +628,10 @@ class InstrumentAPI extends DataSource {
       response.message = 'ERROR: category cannot have white spaces';
       return JSON.stringify(response);
     }
+    if (!this.checkPermission()) {
+      response.message = 'ERROR: User does not have permission.';
+      return JSON.stringify(response);
+    }
     const storeModel = await this.store;
     this.store = storeModel;
     await this.getInstrumentCategory({ name }).then((value) => {
@@ -630,6 +650,10 @@ class InstrumentAPI extends DataSource {
 
   async removeInstrumentCategory({ name }) {
     const response = { message: '', success: false };
+    if (!this.checkPermission()) {
+      response.message = 'ERROR: User does not have permission.';
+      return JSON.stringify(response);
+    }
     const storeModel = await this.store;
     this.store = storeModel;
     await this.getInstrumentCategory({ name }).then((value) => {
@@ -650,6 +674,10 @@ class InstrumentAPI extends DataSource {
 
   async editInstrumentCategory({ currentName, updatedName }) {
     const response = { message: '', success: false };
+    if (!this.checkPermission()) {
+      response.message = 'ERROR: User does not have permission.';
+      return JSON.stringify(response);
+    }
     if (hasWhiteSpace(updatedName)) {
       response.message = 'ERROR: category cannot have white spaces';
       return JSON.stringify(response);
@@ -687,6 +715,10 @@ class InstrumentAPI extends DataSource {
     vendor, modelNumber, serialNumber, category,
   }) {
     const response = { message: '', success: false };
+    if (!this.checkPermission()) {
+      response.message = 'ERROR: User does not have permission.';
+      return JSON.stringify(response);
+    }
     const storeModel = await this.store;
     this.store = storeModel;
     await this.getInstrument({ modelNumber, vendor, serialNumber }).then(async (value) => {
@@ -717,6 +749,10 @@ class InstrumentAPI extends DataSource {
     vendor, modelNumber, serialNumber, category,
   }) {
     const response = { message: '', success: false };
+    if (!this.checkPermission()) {
+      response.message = 'ERROR: User does not have permission.';
+      return JSON.stringify(response);
+    }
     const storeModel = await this.store;
     this.store = storeModel;
     await this.getInstrument({ modelNumber, vendor, serialNumber }).then(async (value) => {
